requirements.txt
streamlit==1.18.0 
Pillow==9.1.0 
requests==2.28.1
duckduckgo-search==0.5.1
